- Created a base_test_case at the project level, so all the apps in this project can re-use the setUpClass.  The setup method will still be app-dependent. - Updated cross_listing_base_test_case to use the setupClass at the project.
